update docs and renaming files
