+++ apps/next/src/content/docs/llamaindex/modules/data/readers/index.mdx
@@ -32,14 +32,6 @@ import { HTMLReader } from '@llamaindex/readers/html'
 
 Additionally the following loaders exist without separate documentation:
 
-- `AssemblyAIReader` transcribes audio using [AssemblyAI](https://www.assemblyai.com/).
-  - [AudioTranscriptReader](/docs/api/classes/AudioTranscriptReader): loads entire transcript as a single document.
-  - [AudioTranscriptParagraphsReader](/docs/api/classes/AudioTranscriptParagraphsReader): creates a document per paragraph.
-  - [AudioTranscriptSentencesReader](/docs/api/classes/AudioTranscriptSentencesReader): creates a document per sentence.
-  - [AudioSubtitlesReader](/docs/api/classes/AudioTranscriptParagraphsReader): creates a document containing the subtitles of a transcript.
-- [NotionReader](/docs/api/classes/NotionReader) loads [Notion](https://www.notion.so/) pages.
-- [SimpleMongoReader](/docs/api/classes/SimpleMongoReader) loads data from a [MongoDB](https://www.mongodb.com/).
-
 Check the [LlamaIndexTS Github](https://github.com/run-llama/LlamaIndexTS) for the most up to date overview of integrations.
 
 ## SimpleDirectoryReader
