+++ packages/providers/discord/src/index.ts
@@ -1,1 +1,1 @@
-export * from "./discord";
+export * from "./reader";
