+++ packages/providers/discord/src/reader.ts
