+++ packages/providers/notion/src/index.ts
@@ -1,1 +1,1 @@
-export * from "./notion";
+export * from "./reader";
